enable-upload-power for ZNCZ01ZM
The power status will only be reported every time the switch is turned on. So this switch is very useful, and it needs to be turned on once by automation, and the status will be reported once.

It would be better if this custom_components can automatically time the operation.